# api/swagger.yaml
openapi: 3.0.1
info:
  title: Product Catalog Management
  description: "## TMF API Reference: TMF620 - Product Catalog Management\n\n### Release\
    \ : 19.0 - June 2019\n\nProduct Catalog API is one of Catalog Management API Family.\
    \ Product Catalog API goal is to provide a catalog of products. \n\n### Operations\n\
    Product Catalog API performs the following operations on the resources :\n- Retrieve\
    \ an entity or a collection of entities depending on filter criteria\n- Partial\
    \ update of an entity (including updating rules)\n- Create an entity (including\
    \ default values and creation rules)\n- Delete an entity\n- Manage notification\
    \ of events"
  version: 4.0.0
servers:
- url: https://serverRoot/tmf-api/productCatalogManagement/v4/
...
paths:
...
  /productSpecification:
    get:
      tags:
      - productSpecification
      summary: List or find ProductSpecification objects
      description: This operation list or find ProductSpecification entities
      operationId: listProductSpecification
      parameters:
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: offset
        in: query
        description: Requested index for start of resources to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: limit
        in: query
        description: Requested number of resources to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          headers:
            X-Total-Count:
              description: Total number of items matching criteria
              style: simple
              explode: false
              schema:
                type: integer
            X-Result-Count:
              description: Actual number of items returned in the response body
              style: simple
              explode: false
              schema:
                type: integer
          content:
            application/json;charset=utf-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductSpecification'
                x-content-type: application/json;charset=utf-8
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: ProductSpecification
    post:
      tags:
      - productSpecification
      summary: Creates a ProductSpecification
      description: This operation creates a ProductSpecification entity.
      operationId: createProductSpecification
      requestBody:
        description: The ProductSpecification to be created
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/ProductSpecification_Create'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ProductSpecification'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: productSpecification
      x-swagger-router-controller: ProductSpecification
  /productSpecification/{id}:
    get:
      tags:
      - productSpecification
      summary: Retrieves a ProductSpecification by ID
      description: This operation retrieves a ProductSpecification entity. Attribute
        selection is enabled for all first level attributes.
      operationId: retrieveProductSpecification
      parameters:
      - name: id
        in: path
        description: Identifier of the ProductSpecification
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: fields
        in: query
        description: Comma-separated properties to provide in response
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Success
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ProductSpecification'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: ProductSpecification
    delete:
      tags:
      - productSpecification
      summary: Deletes a ProductSpecification
      description: This operation deletes a ProductSpecification entity.
      operationId: deleteProductSpecification
      parameters:
      - name: id
        in: path
        description: Identifier of the ProductSpecification
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: Deleted
          content: {}
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: ProductSpecification
    patch:
      tags:
      - productSpecification
      summary: Updates partially a ProductSpecification
      description: This operation updates partially a ProductSpecification entity.
      operationId: patchProductSpecification
      parameters:
      - name: id
        in: path
        description: Identifier of the ProductSpecification
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: The ProductSpecification to be updated
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/ProductSpecification_Update'
        required: true
      responses:
        200:
          description: Updated
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ProductSpecification'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: productSpecification
      x-swagger-router-controller: ProductSpecification
...
components:
  schemas:
...
    AttachmentRefOrValue:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for this particular attachment
        href:
          type: string
          description: URI for this Attachment
        attachmentType:
          type: string
          description: Attachment type such as video, picture
        content:
          type: string
          description: The actual contents of the attachment object, if embedded,
            encoded as base64
        description:
          type: string
          description: A narrative text describing the content of the attachment
        mimeType:
          type: string
          description: Attachment mime type such as extension file for video, picture
            and document
        name:
          type: string
          description: The name of the attachment
        url:
          type: string
          description: Uniform Resource Locator, is a web page address (a subset of
            URI)
        size:
          $ref: '#/components/schemas/Quantity'
        validFor:
          $ref: '#/components/schemas/TimePeriod'
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name
        '@referredType':
          type: string
          description: The actual type of the target instance when needed for disambiguation.
      description: An attachment by value or by reference. For AttachmentRefOrValue,
        the attribute type,schemaLocation and referredType are related to the contained
        entity and not to AttchmentRefOrValue itself
      example:
        '@referredType': '@referredType'
        attachmentType: attachmentType
        validFor:
          startDateTime: 2000-01-23T04:56:07.000+00:00
          endDateTime: 2000-01-23T04:56:07.000+00:00
        '@type': '@type'
        description: description
        mimeType: mimeType
        content: content
        url: url
        size:
          amount: 0.8008282
          units: units
        '@baseType': '@baseType'
        name: name
        id: id
        href: href
        '@schemaLocation': http://example.com/aeiou
...
    ProductSpecification:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the product specification
        href:
          type: string
          description: Reference of the product specification
        brand:
          type: string
          description: The manufacturer or trademark of the specification
        description:
          type: string
          description: A narrative that explains in detail what the product specification
            is
        isBundle:
          type: boolean
          description: isBundle determines whether a productSpecification represents
            a single productSpecification (false), or a bundle of productSpecification
            (true).
        lastUpdate:
          type: string
          description: Date and time of the last update
          format: date-time
        lifecycleStatus:
          type: string
          description: Used to indicate the current lifecycle status
        name:
          type: string
          description: Name of the product specification
        productNumber:
          type: string
          description: An identification number assigned to uniquely identity the
            specification
        version:
          type: string
          description: Product specification version
        attachment:
          type: array
          description: Complements the description of an element (for instance a product)
            through video, pictures...
          items:
            $ref: '#/components/schemas/AttachmentRefOrValue'
        bundledProductSpecification:
          type: array
          description: A type of ProductSpecification that belongs to a grouping of
            ProductSpecifications made available to the market. It inherits of all
            attributes of ProductSpecification.
          items:
            $ref: '#/components/schemas/BundledProductSpecification'
        productSpecCharacteristic:
          type: array
          description: A characteristic quality or distinctive feature of a ProductSpecification.  The
            characteristic can be take on a discrete value, such as color, can take
            on a range of values, (for example, sensitivity of 100-240 mV), or can
            be derived from a formula (for example, usage time (hrs) = 30 - talk time
            *3). Certain characteristics, such as color, may be configured during
            the ordering or some other process.
          items:
            $ref: '#/components/schemas/ProductSpecificationCharacteristic'
        productSpecificationRelationship:
          type: array
          description: A migration, substitution, dependency or exclusivity relationship
            between/among product specifications.
          items:
            $ref: '#/components/schemas/ProductSpecificationRelationship'
        relatedParty:
          type: array
          description: A related party defines party or party role linked to a specific
            entity.
          items:
            $ref: '#/components/schemas/RelatedParty'
        resourceSpecification:
          type: array
          description: The ResourceSpecification is required to realize a ProductSpecification.
          items:
            $ref: '#/components/schemas/ResourceSpecificationRef'
        serviceSpecification:
          type: array
          description: ServiceSpecification(s) required to realize a ProductSpecification.
          items:
            $ref: '#/components/schemas/ServiceSpecificationRef'
        targetProductSchema:
          $ref: '#/components/schemas/TargetProductSchema'
        validFor:
          $ref: '#/components/schemas/TimePeriod'
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name
      description: Is a detailed description of a tangible or intangible object made
        available externally in the form of a ProductOffering to customers or other
        parties playing a party role.
      example:
        isBundle: true
        productSpecCharacteristic:
        - '@type': '@type'
          isUnique: true
          description: description
          maxCardinality: 0
          minCardinality: 6
          regex: regex
          '@baseType': '@baseType'
          valueType: valueType
          name: name
          productSpecCharRelationship:
          - charSpecSeq: 1
            relationshipType: relationshipType
            '@baseType': '@baseType'
            validFor:
              startDateTime: 2000-01-23T04:56:07.000+00:00
              endDateTime: 2000-01-23T04:56:07.000+00:00
            '@type': '@type'
            name: name
            id: id
            href: href
            '@schemaLocation': http://example.com/aeiou
          - charSpecSeq: 1
            relationshipType: relationshipType
            '@baseType': '@baseType'
            validFor:
              startDateTime: 2000-01-23T04:56:07.000+00:00
              endDateTime: 2000-01-23T04:56:07.000+00:00
            '@type': '@type'
            name: name
            id: id
            href: href
            '@schemaLocation': http://example.com/aeiou
          productSpecCharacteristicValue:
          - rangeInterval: rangeInterval
            isDefault: true
            regex: regex
            valueTo: valueTo
            unitOfMeasure: unitOfMeasure
            '@baseType': '@baseType'
            '@type': '@type'
            valueType: valueType
            '@schemaLocation': http://example.com/aeiou
            value: {}
            valueFrom: valueFrom
          - rangeInterval: rangeInterval
            isDefault: true
            regex: regex
            valueTo: valueTo
            unitOfMeasure: unitOfMeasure
            '@baseType': '@baseType'
            '@type': '@type'
            valueType: valueType
            '@schemaLocation': http://example.com/aeiou
            value: {}
            valueFrom: valueFrom
          '@schemaLocation': http://example.com/aeiou
          extensible: true
          configurable: true
        - '@type': '@type'
          isUnique: true
          description: description
          maxCardinality: 0
          minCardinality: 6
          regex: regex
          '@baseType': '@baseType'
          valueType: valueType
          name: name
          productSpecCharRelationship:
          - charSpecSeq: 1
            relationshipType: relationshipType
            '@baseType': '@baseType'
            validFor:
              startDateTime: 2000-01-23T04:56:07.000+00:00
              endDateTime: 2000-01-23T04:56:07.000+00:00
            '@type': '@type'
            name: name
            id: id
            href: href
            '@schemaLocation': http://example.com/aeiou
          - charSpecSeq: 1
            relationshipType: relationshipType
            '@baseType': '@baseType'
            validFor:
              startDateTime: 2000-01-23T04:56:07.000+00:00
              endDateTime: 2000-01-23T04:56:07.000+00:00
            '@type': '@type'
            name: name
            id: id
            href: href
            '@schemaLocation': http://example.com/aeiou
          productSpecCharacteristicValue:
          - rangeInterval: rangeInterval
            isDefault: true
            regex: regex
            valueTo: valueTo
            unitOfMeasure: unitOfMeasure
            '@baseType': '@baseType'
            '@type': '@type'
            valueType: valueType
            '@schemaLocation': http://example.com/aeiou
            value: {}
            valueFrom: valueFrom
          - rangeInterval: rangeInterval
            isDefault: true
            regex: regex
            valueTo: valueTo
            unitOfMeasure: unitOfMeasure
            '@baseType': '@baseType'
            '@type': '@type'
            valueType: valueType
            '@schemaLocation': http://example.com/aeiou
            value: {}
            valueFrom: valueFrom
          '@schemaLocation': http://example.com/aeiou
          extensible: true
          configurable: true
        lifecycleStatus: lifecycleStatus
        targetProductSchema:
          '@baseType': '@baseType'
          '@type': '@type'
          '@schemaLocation': '@schemaLocation'
        '@type': '@type'
        resourceSpecification:
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
          version: version
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
          version: version
        serviceSpecification:
        - targetServiceSchema:
            '@baseType': '@baseType'
            '@type': '@type'
            '@schemaLocation': '@schemaLocation'
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
          version: version
        - targetServiceSchema:
            '@baseType': '@baseType'
            '@type': '@type'
            '@schemaLocation': '@schemaLocation'
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
          version: version
        description: description
        productNumber: productNumber
        relatedParty:
        - '@referredType': '@referredType'
          role: role
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        - '@referredType': '@referredType'
          role: role
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        version: version
        attachment:
        - '@referredType': '@referredType'
          attachmentType: attachmentType
          validFor:
            startDateTime: 2000-01-23T04:56:07.000+00:00
            endDateTime: 2000-01-23T04:56:07.000+00:00
          '@type': '@type'
          description: description
          mimeType: mimeType
          content: content
          url: url
          size:
            amount: 0.8008282
            units: units
          '@baseType': '@baseType'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        - '@referredType': '@referredType'
          attachmentType: attachmentType
          validFor:
            startDateTime: 2000-01-23T04:56:07.000+00:00
            endDateTime: 2000-01-23T04:56:07.000+00:00
          '@type': '@type'
          description: description
          mimeType: mimeType
          content: content
          url: url
          size:
            amount: 0.8008282
            units: units
          '@baseType': '@baseType'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        '@baseType': '@baseType'
        bundledProductSpecification:
        - lifecycleStatus: lifecycleStatus
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        - lifecycleStatus: lifecycleStatus
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        lastUpdate: 2000-01-23T04:56:07.000+00:00
        name: name
        productSpecificationRelationship:
        - relationshipType: relationshipType
          '@baseType': '@baseType'
          '@type': '@type'
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        - relationshipType: relationshipType
          '@baseType': '@baseType'
          '@type': '@type'
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        id: id
        href: href
        '@schemaLocation': http://example.com/aeiou
        brand: brand
    ProductSpecification_Create:
      required:
      - name
      type: object
      properties:
        brand:
          type: string
          description: The manufacturer or trademark of the specification
        description:
          type: string
          description: A narrative that explains in detail what the product specification
            is
        isBundle:
          type: boolean
          description: isBundle determines whether a productSpecification represents
            a single productSpecification (false), or a bundle of productSpecification
            (true).
        lastUpdate:
          type: string
          description: Date and time of the last update
          format: date-time
        lifecycleStatus:
          type: string
          description: Used to indicate the current lifecycle status
        name:
          type: string
          description: Name of the product specification
        productNumber:
          type: string
          description: An identification number assigned to uniquely identity the
            specification
        version:
          type: string
          description: Product specification version
        attachment:
          type: array
          description: Complements the description of an element (for instance a product)
            through video, pictures...
          items:
            $ref: '#/components/schemas/AttachmentRefOrValue'
        bundledProductSpecification:
          type: array
          description: A type of ProductSpecification that belongs to a grouping of
            ProductSpecifications made available to the market. It inherits of all
            attributes of ProductSpecification.
          items:
            $ref: '#/components/schemas/BundledProductSpecification'
        productSpecCharacteristic:
          type: array
          description: A characteristic quality or distinctive feature of a ProductSpecification.  The
            characteristic can be take on a discrete value, such as color, can take
            on a range of values, (for example, sensitivity of 100-240 mV), or can
            be derived from a formula (for example, usage time (hrs) = 30 - talk time
            *3). Certain characteristics, such as color, may be configured during
            the ordering or some other process.
          items:
            $ref: '#/components/schemas/ProductSpecificationCharacteristic'
        productSpecificationRelationship:
          type: array
          description: A migration, substitution, dependency or exclusivity relationship
            between/among product specifications.
          items:
            $ref: '#/components/schemas/ProductSpecificationRelationship'
        relatedParty:
          type: array
          description: A related party defines party or party role linked to a specific
            entity.
          items:
            $ref: '#/components/schemas/RelatedParty'
        resourceSpecification:
          type: array
          description: The ResourceSpecification is required to realize a ProductSpecification.
          items:
            $ref: '#/components/schemas/ResourceSpecificationRef'
        serviceSpecification:
          type: array
          description: ServiceSpecification(s) required to realize a ProductSpecification.
          items:
            $ref: '#/components/schemas/ServiceSpecificationRef'
        targetProductSchema:
          $ref: '#/components/schemas/TargetProductSchema'
        validFor:
          $ref: '#/components/schemas/TimePeriod'
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name
      description: |-
        Is a detailed description of a tangible or intangible object made available externally in the form of a ProductOffering to customers or other parties playing a party role.
        Skipped properties: id,href
...
    TimePeriod:
      type: object
      properties:
        endDateTime:
          type: string
          description: End of the time period, using IETC-RFC-3339 format
          format: date-time
        startDateTime:
          type: string
          description: Start of the time period, using IETC-RFC-3339 format. If you
            define a start, you must also define an end
          format: date-time
      description: A period of time, either as a deadline (endDateTime only) a startDateTime
        only, or both
      example:
        startDateTime: 2000-01-23T04:56:07.000+00:00
        endDateTime: 2000-01-23T04:56:07.000+00:00
...
    Error:
      required:
      - code
      - reason
      type: object
      properties:
        code:
          type: string
          description: Application relevant detail, defined in the API or a common
            list.
        reason:
          type: string
          description: Explanation of the reason for the error which can be shown
            to a client user.
        message:
          type: string
          description: More details and corrective actions related to the error which
            can be shown to a client user.
        status:
          type: string
          description: HTTP Error code extension
        referenceError:
          type: string
          description: URI of documentation describing the error.
          format: uri
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class.
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name.
      description: Used when an API throws an Error, typically with a HTTP error response-code
        (3xx, 4xx, 5xx)